Lesson3 and 4 updated, classes renamed, empty strings deleted
