+++ Lesson2/src/main/java/Test.java
@@ -1,6 +1,5 @@
 public class Test {
     public static void main(String[] args) {
         System.out.println("Hello git");
-
     }
 }
