+++ Lesson2/src/main/java/Test3.java
@@ -1,5 +1,4 @@
 public class Test3 {
     public static void main(String[] args) {
-
     }
 }
