+++ Lesson3/src/main/java/CountMaxNumberFromArray.java
@@ -1,9 +1,8 @@
-public class countMaxNumberFromArray {
+public class CountMaxNumberFromArray {
     public static void main(String[] args) {
         int[] nums = {1, 2, 10, 3};
         int x = max(nums);
         System.out.println(x);
-
     }
 
     public static int max(int[] nums) {
