+++ Lesson3/src/main/java/CountMaxValue.java
@@ -1,0 +1,15 @@
+public class CountMaxValue {
+    public static void main(String[] args) {
+        int a = 4;//10;
+        int b = 5;//10;
+        int x = max(a, b);
+        System.out.println(x);
+    }
+
+    public static int max(int a, int b) {
+        return Math.max(a, b);
+    }
+
+}
+
+
