+++ Lesson4/src/main/java/CountAlphabet.java
@@ -1,8 +1,6 @@
-public class countAlphabet1 {
+public class CountAlphabet {
     public static void main(String[] args) {
-
         for (char a = 'a'; a <= 'z'; a++) {
-
             System.out.print(" " + a);
         }
     }
