+++ Lesson4/src/main/java/ShowChart.java
@@ -1,18 +1,16 @@
 
-public class showChart5 {
+public class ShowChart {
     public static void main(String[] args) {
         String s1 = "      \tbyte \tshort  \tchar  \tint  \tlong  \tfloat  \tdouble \tboolean";
         System.out.println(s1);
-
-        String s2 = "byte   \tt   \tня     \tx     \tня   \tня    \tня     \tня     \tx";
+        String s2 = "byte   \tt   \tня     \tя     \tня   \tня    \tя      \tня     \tx";
         String s3 = "short  \tя   \tt      \tя     \tня   \tня    \tя      \tня     \tx";
-        String s4 = "char   \tя   \tя      \tt     \tня   \tня    \tя      \tня     \tx";
+        String s4 = "char   \tя   \tя      \tt     \tня   \tня    \tя      \tя      \tx";
         String s5 = "int    \tя   \tя      \tя     \tt    \tня    \tя      \tня     \tx";
         String s6 = "long   \tя   \tя      \tя     \tя    \tt     \tя      \tя      \tx";
         String s7 = "float  \tя   \tя      \tя     \tя    \tя     \tt      \tня     \tх";
         String s8 = "double \tя   \tя      \tя     \tя    \tя     \tя      \tt      \tx";
         String s9 = "boolean\tx   \tx      \tx     \tx    \tx     \tx      \tx      \tt";
         System.out.println(s2 + "\n" + s3 + "\n" + s4 + "\n" + s5 + "\n" + s6 + "\n" + s7 + "\n" + s8 + "\n" + s9);
-
     }
 }
